Honor SkillsBench build stall timeout
