+++ scripts/skillsbench_runner_constants.py
@@ -48,7 +48,6 @@ DEFAULT_APP_SERVER_GOAL_FIRST_ACTION_TIMEOUT_SEC = 3600
 DEFAULT_CODEX_CLI_GOAL_FIRST_ACTION_TIMEOUT_SEC = 900
 DEFAULT_CODEX_CLI_GOAL_ACTIVE_TIMEOUT_SEC = 180
 DEFAULT_BUILD_STALL_TIMEOUT_SEC = 900
-MAX_BUILD_STALL_TIMEOUT_SEC = 900
 DEFAULT_VERIFIER_PREP_TIMEOUT_SEC = 120
 DEFAULT_SOFT_VERIFIER_TIMEOUT_SEC = 600
 DEFAULT_PRODUCT_MODE_SOFT_VERIFY_POLICY = "every-round"
